# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Informations",
        "contact": {
            "email": "skpandit659@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=CarInformations"
        }
    },
    "paths": {
        "/api/car": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get All cars",
                "operationId": "all",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Your API Key",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation Successful"
                    },
                    "403": {
                        "description": "Invalid API key"
                    }
                }
            }
        },
        "/api/car/{id}": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Get a car by ID",
                "operationId": "findCarById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the car to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Your API Key",
                        "required": true
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Invalid API key"
                    },
                    "404": {
                        "description": "Resource not found for id: {id}"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Car"
        }
    ]
}